Remove transmutes that rely on unstable layout
The previous transmutes in this crate relied on the layout of slices,
and also on the layout of repr(Rust) types in this crate. This changes
all the transmutes to fat pointer casts, which do not rely on the layout
of slices, and which are valid because this adds repr(transparent) to
IoVec.